Renaming namespace `global::ShellFileDialogs` to `global::MvvmDialogs.ComShellDialogs`
