+++ src/net/ComShellDialogs/Native/Enums.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace ShellFileDialogs
+namespace MvvmDialogs.ComShellDialogs
 {
 	internal enum SICHINTF
 	{
