+++ src/net/ComShellDialogs/Native/FileOpenDialogRCW.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Runtime.InteropServices;
 
-namespace ShellFileDialogs
+namespace MvvmDialogs.ComShellDialogs
 {
 	[ComImport]
 	[ClassInterface( ClassInterfaceType.None )]
