+++ src/net/ComShellDialogs/Native/IFileDialog.cs
@@ -2,7 +2,7 @@
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
-namespace ShellFileDialogs
+namespace MvvmDialogs.ComShellDialogs
 {
 	// Disable warning if a method declaration hides another inherited from a parent COM interface
 	// To successfully import a COM interface, all inherited methods need to be declared again with
