+++ src/net/ComShellDialogs/Native/IFileDialogEvents.cs
@@ -2,7 +2,7 @@
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
-namespace ShellFileDialogs
+namespace MvvmDialogs.ComShellDialogs
 {
 	[ComImport]
 	[Guid( ShellIIDGuid.IFileDialogEvents )]
