+++ src/net/ComShellDialogs/Native/IModalWindow.cs
@@ -2,7 +2,7 @@
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
-namespace ShellFileDialogs
+namespace MvvmDialogs.ComShellDialogs
 {
 	[ComImport]
 	[Guid( ShellIIDGuid.IModalWindow )]
