+++ src/net/ComShellDialogs/Native/IShellItem2.cs
@@ -3,7 +3,7 @@ using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using System.Runtime.InteropServices.ComTypes;
 
-namespace ShellFileDialogs
+namespace MvvmDialogs.ComShellDialogs
 {
 	// Disable warning if a method declaration hides another inherited from a parent COM interface
 	// To successfully import a COM interface, all inherited methods need to be declared again with
