+++ src/net/ComShellDialogs/Native/Identifiers.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace ShellFileDialogs
+namespace MvvmDialogs.ComShellDialogs
 {
 	internal static class ShellIIDGuid
 	{
