+++ src/net/ComShellDialogs/Native/NativeMethods.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Runtime.InteropServices;
 
-namespace ShellFileDialogs
+namespace MvvmDialogs.ComShellDialogs
 {
 	internal static class ShellNativeMethods
 	{
