+++ src/net/ComShellDialogs/Native/Properties/Enums.cs
@@ -2,7 +2,7 @@
 
 using System;
 
-namespace ShellFileDialogs
+namespace MvvmDialogs.ComShellDialogs
 {
 	internal enum RelativeDescriptionType
 	{
