+++ src/net/ComShellDialogs/Native/Properties/IPropertyStore.cs
@@ -4,7 +4,7 @@ using System;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
-namespace ShellFileDialogs
+namespace MvvmDialogs.ComShellDialogs
 {
 	/// <summary>
 	/// A property store
