+++ src/net/ComShellDialogs/Native/Properties/PropVariant.cs
@@ -6,7 +6,7 @@ using System.Linq.Expressions;
 using System.Reflection;
 using System.Runtime.InteropServices;
 
-namespace ShellFileDialogs
+namespace MvvmDialogs.ComShellDialogs
 {
 	/// <summary>
 	/// Represents the OLE struct PROPVARIANT.
