+++ src/net/ComShellDialogs/Native/Properties/PropVariantNativeMethods.cs
@@ -5,7 +5,7 @@
 using System;
 using System.Runtime.InteropServices;
 
-namespace ShellFileDialogs
+namespace MvvmDialogs.ComShellDialogs
 {
 	internal static class PropVariantNativeMethods
 	{
